Refactor satellite dataset scripts
