+++ scripts/validate_satellite_dataset.py
@@ -1,21 +1,21 @@
 from argparse import ArgumentParser, Namespace
 from pathlib import Path
 
-import source.dataset_validation as validation
-import source.utility as util
+import source.satellite_dataset.utility as sd_util
+import source.satellite_dataset.validation as sd_validation
 
 
 def main() -> None:
     input_args = parse_input_args()
     dataset_name: str | None = input_args.name
 
-    _, dataset = util.load_satellite_dataset(dataset_name=dataset_name)
+    _, dataset = sd_util.load_satellite_dataset(dataset_name=dataset_name)
     dataset_archive = dataset["archive"]
     dataset_path = Path(dataset["path"])
 
     print()  # Empty line for better separation
 
-    is_valid, message = validation.validate_satellite_dataset(
+    is_valid, message = sd_validation.validate_satellite_dataset(
         dataset_archive, dataset_path
     )
